+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - perks managing for Fallout characters
 - dice rolls for skills checks for Fallout characters
+- equipment management for Fallout characters
 
 ## [0.4.13] - 2026-03-04
 ### Modified